Add an agent detail page.
This PR adds a new page that displays details of a particular supplement
or drug. This in turn allows us to truncate the information we display
on the results page, so that results can be scanned more easily.

While doing this I also split out the components so that they could
be easily reused across different pages.

* Fixes #15; Add the info icon and expose the definition when it's
  hovered.
* Fixes #16; Use a search input, so that searching only occurs
  when enter is pressed or the search button is clicked.
* Fixes #18; Truncate the information we display on the results page.
* Fixes #13; Add a supplement detail page.