Style disclaimer (#41)
This fixes #19, by styling the disclaimer per mocks and
adding a dismiss button. The dismiss button sets a cookie. When
the cookie is set the disclaimer is no longer rendered.

I also fixed a small UI nit, where the footer isn't anchored
to the bottom of the page if the content is smaller than the
viewport.