Remove redirect route.
The handlers for routes `/a/:slug/:id` and `/a/:id` were both
being invoked when a client requested a URL like `/a/foo/`, where
`foo` could be either the `id` or `slug`.  This would cause a `500`
when the provided value wasn't a valid `CUI`.

To fix this I simply removed the `/a/:id` path, given it's not
officially removed.