API docs + a fix for the remote_addr.
We're going to advertise the public API to assess adoption. This
adds some controls for composing API documentation, and docs for
the search endpoint.

We plan to add more endpoints on a case by case basis, and then
add a link to the home page.

This fix also includes a change that fixes the value of `remote_addr` so 
that it's the actual client IP, not that of the load balancer.