Only display 3 pieces of evidence on the agent detail page.
The interaction detail pages aren't being indexed by Google, even
though these are highly searchable.

We think it's because they look like duplicate content, since all
mentions are viewable on both the agent and interaction detail pages.
This modifies that so that only 3 are displayed on the agent detail page,
while all are visible on the interaction detail page.