Fix a CVE.
For `serialize-javascript` to be updated, using Yarn's `resolutions`
mechanism.

Sadly, `NextJS` hasn't updated to not include the dependency with an
issue yet, so we have to force a transitive update. While in theory
this is risky, I tested everything to make sure it's working. We've
also done this in several other places (like the AI2 website), so
we should feel pretty confident.