Force `node-fetch` to the latest.
This project depends on an old version of both `antd` and
`@allenai/varnish`, both of which rely on a dependencies which
eventally require vulernable versions of `node-fetch`.

Rather than play dependency whack-a-mole more I'm taking the
nuclear route and just force updating this to the latest version.
This does in one case for a potentially non-backwards compatible change,
from `1.*` to `2.*`, but things appeared to work fine (they didn't
on the Leaderboard), so we can call this good.

I do worry about maintaining a bunch of `resolutions` as time goes
on, but perhaps that's just the game we're playing if we intend to
keep writing UIs that depend heavily on JavaScript.

https://github.com/allenai/reviz/issues/135