Update ini.
This uses the `resolutions` feature to force `ini` to a version
without the associated CVE. The update should be safe, since it's
a only a patch.

I tried to use the `--ignore-optional` flag `yarn` provides, but
it didn't work (`ini` is required by `rc` which is required by `sharp`
which is an optional dependency of `next`). It appears this is a known
issue: https://github.com/yarnpkg/yarn/issues/5366

Resolves:
https://github.com/allenai/supp.ai/security/dependabot/ui/yarn.lock/ini/open

In relation to:
https://github.com/allenai/reviz/issues/165