Update NextJS (and a bunch of other things)
This applies a bunch of dependency updates, mainly to close
up a few CVEs.

The biggest motivator for these changes is a CVE with NextJS's
image transformation API, which is why we both update Next and
add rules to the proxy that prevent that endpoint from being
accessible.

I was able to get the site up and working with these changes.
I clicked around and everything looked good, so I think this
is safe.

https://github.com/allenai/reviz/issues/402